Make the --dry-run flag do the right thing

+++ vup/__init__.py
@@ -19,16 +19,18 @@ REGEX = (
 
 
 class VersionFile():
-    def __init__(self, filename):
+    def __init__(self, filename, is_dry_run):
         self.filename = filename
+        self.is_dry_run = is_dry_run
         self.version = None
         self._get_version()
 
     def replace_version(self, new_version):
         filedata_to_write = re.sub(
             REGEX, str(new_version), self.filedata, count=1)
-        with open(self.filename, 'w') as a_file:
-            a_file.write(filedata_to_write)
+        if not self.is_dry_run:
+            with open(self.filename, 'w') as a_file:
+                a_file.write(filedata_to_write)
         self._get_version()
 
     def _get_version(self):
@@ -40,17 +42,17 @@ class VersionFile():
         self.version = semantic_version.Version(found_version)
 
 
-def get_bumped_version(version, type_to_bump):
+def get_bumped_version(version, bump_type):
     bumped_version = None
-    if type_to_bump == 'major':
+    if bump_type == 'major':
         bumped_version = version.next_major()
-    elif type_to_bump == 'minor':
+    elif bump_type == 'minor':
         bumped_version = version.next_minor()
-    elif type_to_bump == 'patch':
+    elif bump_type == 'patch':
         if not version.prerelease:
             bumped_version = version.next_patch()
     else:
-        assert False  # 'type_to_bump is invalid'
+        assert False  # 'bump_type is invalid'
     bumped_version.prerelease = None
     return bumped_version
 
@@ -61,28 +63,39 @@ def get_bumped_prerelease_version(version):
     return prerelease_version
 
 
-def commit_version_file_change(repo, filename, old_version, new_version):
-    repo.index.add([filename])
+def commit_version_file_change(repo, filename, old_version, new_version,
+                               is_dry_run):
+    if not is_dry_run:
+        repo.index.add([filename])
     commit_message = 'Increment version from {old_version} to {new_version}'
     commit_message = commit_message.format(
         new_version=new_version, old_version=old_version)
     print(commit_message)
-    repo.index.commit(commit_message)
+    if not is_dry_run:
+        repo.index.commit(commit_message)
 
 
-def tag_version_file_change(repo, version):
+def tag_version_file_change(repo, version, is_dry_run):
     tag_message = 'Version {version}'.format(version=version)
-    repo.create_tag(version, message=tag_message)
+    if not is_dry_run:
+        repo.create_tag(version, message=tag_message)
 
 
-def run_hook(cmd):
-    result = subprocess.run(cmd, shell=True)
-    if result.stdout:
-        print(result.stdout)
-    return result.returncode == 0
+def run_hook(cmd, is_dry_run):
+    print(cmd)
+    if not is_dry_run:
+        result = subprocess.run(cmd, shell=True)
+        if result.stdout:
+            print(result.stdout)
+        return result.returncode == 0
+    return True
 
 
-def bump(filename, bump_type='patch', pre_bump_hook=None, post_bump_hook=None):
+def bump(filename,
+         bump_type='patch',
+         pre_bump_hook=None,
+         post_bump_hook=None,
+         is_dry_run=False):
 
     # TODO handle exceptions if git repo doesn't exist / has issues
     repo = git.Repo('.')
@@ -94,12 +107,12 @@ def bump(filename, bump_type='patch', pre_bump_hook=None, post_bump_hook=None):
     assert is_file_in_repo(repo, os.path.abspath(filename))
 
     if pre_bump_hook:
-        assert run_hook(pre_bump_hook)  # pre_bump hook failed
+        assert run_hook(pre_bump_hook, is_dry_run)  # pre_bump hook failed
 
-    version_file = VersionFile(filename)
+    version_file = VersionFile(filename, is_dry_run)
 
     starting_version = version_file.version
-    release_version = get_bumped_version(version_file.version, type_to_bump)
+    release_version = get_bumped_version(version_file.version, bump_type)
     prerelease_version = get_bumped_prerelease_version(release_version)
 
     # TODO test tag already exists
@@ -107,14 +120,14 @@ def bump(filename, bump_type='patch', pre_bump_hook=None, post_bump_hook=None):
 
     version_file.replace_version(release_version)
     commit_version_file_change(repo, filename, starting_version,
-                               release_version)
-    tag_version_file_change(repo, version_file.version)
+                               release_version, is_dry_run)
+    tag_version_file_change(repo, version_file.version, is_dry_run)
 
     version_file.replace_version(prerelease_version)
     commit_version_file_change(repo, filename, release_version,
-                               prerelease_version)
+                               prerelease_version, is_dry_run)
     if post_bump_hook:
-        assert run_hook(post_bump_hook)  # pre_bump hook failed
+        assert run_hook(post_bump_hook, is_dry_run)  # pre_bump hook failed
 
 
 def is_file_in_repo(repo, a_file):
